requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
